Correccion al modal de ventas en el apartado del cambio v3

+++ src/main/java/com/uriel/pos/sistema_punto_venta/Venta.java
@@ -1,6 +1,5 @@
 package com.uriel.pos.sistema_punto_venta;
 
-import java.math.BigDecimal;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
@@ -35,7 +34,7 @@ public class Venta {
                 data = new ProductoData(
                     rs.getInt("id_producto"),
                     rs.getString("nombre"),
-                    rs.getBigDecimal("precio_venta"),
+                    rs.getDouble("precio_venta"),
                     rs.getInt("existencias_act"),
                     fotoBase64,
                     rs.getString("codigo_barras")
@@ -103,7 +102,7 @@ public class Venta {
                 lista.add(new ProductoData(
                     rs.getInt("id_producto"),
                     rs.getString("nombre"),
-                    rs.getBigDecimal("precio_venta"),
+                    rs.getDouble("precio_venta"),
                     rs.getInt("existencias_act"),
                     fotoBase64,
                     rs.getString("codigo_barras")
@@ -119,7 +118,7 @@ public class Venta {
         return lista;
     }
 
-    public void finalizarTransaccion(int idVenta, List<DetalleVentaRequest> detalles, BigDecimal totalFinal) throws SQLException {
+    public void finalizarTransaccion(int idVenta, List<DetalleVentaRequest> detalles, double totalFinal) throws SQLException {
         Connection c = null;
         try {
             Conexion con = new Conexion();
@@ -134,7 +133,7 @@ public class Venta {
                 String sqlInsertVenta = "INSERT INTO ventas (id_usuario, total, fecha) VALUES (?, ?, CONVERT_TZ(NOW(), '+02:00', '-04:00'))";
                 PreparedStatement stmtVenta = c.prepareStatement(sqlInsertVenta, PreparedStatement.RETURN_GENERATED_KEYS);
                 stmtVenta.setInt(1, 1); // usuario por defecto
-                stmtVenta.setBigDecimal(2, totalFinal);
+                stmtVenta.setDouble(2, totalFinal);
                 stmtVenta.executeUpdate();
                 
                 ResultSet rsVenta = stmtVenta.getGeneratedKeys();
@@ -148,7 +147,7 @@ public class Venta {
 
             String sqlTotal = "UPDATE ventas SET total = ? WHERE id_venta = ?";
             PreparedStatement stmtTotal = c.prepareStatement(sqlTotal);
-            stmtTotal.setBigDecimal(1, totalFinal);
+            stmtTotal.setDouble(1, totalFinal);
             stmtTotal.setInt(2, ventaId);
             int rowsUpdated = stmtTotal.executeUpdate();
             System.out.println("Filas actualizadas en ventas: " + rowsUpdated);
@@ -159,8 +158,8 @@ public class Venta {
                 PreparedStatement stmtInsert = c.prepareStatement(sqlInsert);
                 stmtInsert.setInt(1, ventaId);
                 stmtInsert.setInt(2, detalle.idProducto);
-                stmtInsert.setBigDecimal(3, detalle.cantidad);
-                stmtInsert.setBigDecimal(4, detalle.precioUnitario);
+                stmtInsert.setDouble(3, detalle.cantidad);
+                stmtInsert.setDouble(4, detalle.precioUnitario);
                 stmtInsert.executeUpdate();
                 stmtInsert.close();
 
@@ -194,7 +193,7 @@ public class Venta {
     }
     
     // Guardar venta en espera
-    public int guardarVentaEnEspera(int idUsuario, BigDecimal total, List<DetalleVentaEnEspera> detalles) {
+    public int guardarVentaEnEspera(int idUsuario, double total, List<DetalleVentaEnEspera> detalles) {
         int idVentaEspera = -1;
         try {
             Conexion con = new Conexion();
@@ -203,7 +202,7 @@ public class Venta {
             String sql = "INSERT INTO ventas_en_espera (id_usuario, total, fecha) VALUES (?, ?, CONVERT_TZ(NOW(), '+02:00', '-04:00'))";
             PreparedStatement stmt = c.prepareStatement(sql, PreparedStatement.RETURN_GENERATED_KEYS);
             stmt.setInt(1, idUsuario);
-            stmt.setBigDecimal(2, total);
+            stmt.setDouble(2, total);
             stmt.executeUpdate();
             
             ResultSet rs = stmt.getGeneratedKeys();
@@ -221,8 +220,8 @@ public class Venta {
                 stmtDetalle.setInt(2, detalle.idProducto);
                 stmtDetalle.setString(3, detalle.codigo);
                 stmtDetalle.setString(4, detalle.nombre);
-                stmtDetalle.setBigDecimal(5, detalle.precio);
-                stmtDetalle.setBigDecimal(6, detalle.cantidad);
+                stmtDetalle.setDouble(5, detalle.precio);
+                stmtDetalle.setDouble(6, detalle.cantidad);
                 stmtDetalle.setString(7, detalle.fotoProducto);
                 stmtDetalle.executeUpdate();
                 stmtDetalle.close();
@@ -267,7 +266,7 @@ public class Venta {
                     rs.getInt("id_usuario"),
                     "Usuario #" + rs.getInt("id_usuario"),  // Username fijo por ahora
                     rs.getString("fecha"),
-                    rs.getBigDecimal("total")
+                    rs.getDouble("total")
                 );
                 lista.add(venta);
             }
@@ -291,8 +290,8 @@ public class Venta {
                         rsDetalles.getInt("id_producto"),
                         rsDetalles.getString("codigo"),
                         rsDetalles.getString("nombre"),
-                        rsDetalles.getBigDecimal("precio"),
-                        rsDetalles.getBigDecimal("cantidad"),
+                        rsDetalles.getDouble("precio"),
+                        rsDetalles.getDouble("cantidad"),
                         rsDetalles.getString("foto_producto")
                     );
                     venta.detalles.add(detalle);
@@ -351,8 +350,8 @@ public class Venta {
                         rsDetalles.getInt("id_producto"),
                         rsDetalles.getString("codigo"),
                         rsDetalles.getString("nombre"),
-                        rsDetalles.getBigDecimal("precio"),
-                        rsDetalles.getBigDecimal("cantidad"),
+                        rsDetalles.getDouble("precio"),
+                        rsDetalles.getDouble("cantidad"),
                         rsDetalles.getString("foto_producto")
                     );
                     venta.detalles.add(detalle);
@@ -402,12 +401,12 @@ public class Venta {
 class ProductoData {
     public int idProducto;
     public String nombre;
-    public BigDecimal precioVenta;
+    public double precioVenta;
     public int stockActual;
     public String fotoProducto;
     public String codigoBarras;
 
-    public ProductoData(int idProducto, String nombre, BigDecimal precioVenta, int stockActual, String fotoProducto, String codigoBarras) {
+    public ProductoData(int idProducto, String nombre, double precioVenta, int stockActual, String fotoProducto, String codigoBarras) {
         this.idProducto = idProducto;
         this.nombre = nombre;
         this.precioVenta = precioVenta;
@@ -419,8 +418,8 @@ class ProductoData {
 
 class DetalleVentaRequest {
     public int idProducto;
-    public BigDecimal cantidad;
-    public BigDecimal precioUnitario;
+    public double cantidad;
+    public double precioUnitario;
 }
 
 // Clase para venta en espera
@@ -429,10 +428,10 @@ class VentaEnEspera {
     public int idUsuario;
     public String nombreUsuario;
     public String fecha;
-    public BigDecimal total;
+    public double total;
     public List<DetalleVentaEnEspera> detalles;
 
-    public VentaEnEspera(int id, int idUsuario, String nombreUsuario, String fecha, BigDecimal total) {
+    public VentaEnEspera(int id, int idUsuario, String nombreUsuario, String fecha, double total) {
         this.id = id;
         this.idUsuario = idUsuario;
         this.nombreUsuario = nombreUsuario;
@@ -446,11 +445,11 @@ class DetalleVentaEnEspera {
     public int idProducto;
     public String codigo;
     public String nombre;
-    public BigDecimal precio;
-    public BigDecimal cantidad;
+    public double precio;
+    public double cantidad;
     public String fotoProducto;
 
-    public DetalleVentaEnEspera(int idProducto, String codigo, String nombre, BigDecimal precio, BigDecimal cantidad, String fotoProducto) {
+    public DetalleVentaEnEspera(int idProducto, String codigo, String nombre, double precio, double cantidad, String fotoProducto) {
         this.idProducto = idProducto;
         this.codigo = codigo;
         this.nombre = nombre;
